+++ mk/crates.mk
@@ -68,15 +68,15 @@ DEPS_rustdoc := rustc native:sundown serialize sync getopts collections \
 DEPS_flate := std native:miniz
 DEPS_arena := std collections
 DEPS_glob := std
-DEPS_serialize := std
+DEPS_serialize := std collections
 DEPS_term := std collections
 DEPS_semver := std
 DEPS_uuid := std serialize
 DEPS_sync := std
 DEPS_getopts := std
-DEPS_collections := std serialize
+DEPS_collections := std
 DEPS_fourcc := syntax std
-DEPS_num := std extra
+DEPS_num := std
 DEPS_test := std extra collections getopts serialize term
 DEPS_time := std serialize
