+++ src/libnum/lib.rs
@@ -15,8 +15,6 @@
 #[crate_type = "dylib"];
 #[license = "MIT/ASL2"];
 
-extern crate extra;
-
 pub mod bigint;
 pub mod rational;
 pub mod complex;
